* Merge in updates from kabanero collections

* Minor fixes to catch things missed in the merge.

+++ ci/ext/add_collection_resources.sh
@@ -123,7 +123,6 @@ then
         if [ $default_imageId == $imageId ]
         then
             default_image=$(yq r $index_file images.[$count].image)
-            default_image="${default_image/\$IMAGE_REGISTRY_ORG/${IMAGE_REGISTRY_ORG}}"
         fi
         count=$(( $count + 1 ))
     done
@@ -141,17 +140,19 @@ then
             mkdir -p $template_temp
 
             # Update template archives
-            tar -xzf $assets_dir/$template_archive -C $template_temp
-            if [ -f $template_temp/.appsody-config.yaml ]
-            then 
-                yq w -i $template_temp/.appsody-config.yaml stack $default_image 
-            else
-                echo "stack: $default_image" > $template_temp/.appsody-config.yaml
-            fi
-            tar -czf $assets_dir/$template_archive -C $template_temp .
-            echo -e "--- Updated template archive: $template_archive"
+            if [ -f $assets_dir/$template_archive ]; then
+                tar -xzf $assets_dir/$template_archive -C $template_temp
+                if [ -f $template_temp/.appsody-config.yaml ]
+                then 
+                    yq w -i $template_temp/.appsody-config.yaml stack $default_image 
+                else
+                    echo "stack: $default_image" > $template_temp/.appsody-config.yaml
+                fi
+                tar -czf $assets_dir/$template_archive -C $template_temp .
+                echo -e "--- Updated template archive: $template_archive"
         
-            rm -fr $template_temp
+                rm -fr $template_temp
+            fi
         fi
     done
 fi
